+++ Demo/Sources/Demos/Demo Screens/ReorderingViewController.swift
@@ -16,14 +16,10 @@ import BlueprintUICommonControls
 
 final class ReorderingViewController : ListViewController
 {
-    fileprivate var reorderingStorage: ReorderingStorage!
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Reload", style: .plain, target: self, action: #selector(reload))
-        
-        reorderingStorage = .init()
     }
     
     @objc private func reload() {
@@ -35,15 +31,12 @@ final class ReorderingViewController : ListViewController
         list.appearance = .demoAppearance
         list.layout = .demoLayout
         
-        list.stateObserver.onItemReordered { [weak self] reordered in
+        list.stateObserver.onItemReordered { reordered in
             print("Moved: \(reordered.result.indexPathsDescription)")
             
             reordered.result.toSection.filtered(to: DemoItem.self) { items in
                 print(items.map(\.text).joined(separator: "\n"))
             }
-            
-            self?.reorderingStorage.reset()
-            self?.reload()
         }
         
         list += Section("1") { section in
@@ -103,48 +96,17 @@ final class ReorderingViewController : ListViewController
         list += Section("4") { section in
             section.header = DemoHeader(title: "Long press")
             
-            reorderingStorage.rows.forEach {
-                section += Item($0) { item in
-                    item.reordering = ItemReordering(sections: .current)
-                    item.onStartReorder = { [weak self] item in
-                        self?.reorderingStorage.setDragging(item.identifier.value, dragging: true)
-                        DispatchQueue.main.async {
-                            self?.reload()
-                        }
-                    }
-                }
+            section += Item(DemoItem(text: "3,0 Row (long press)", requiresLongPress: true)) { item in
+                item.reordering = ItemReordering(sections: .current)
             }
-        }
-    }
-}
 
-private final class ReorderingStorage {
-    
-    var rows: [DemoItem] {
-        allRows.values.sorted {
-            $0.identifierValue < $1.identifierValue
-        }
-    }
+            section += Item(DemoItem(text: "3,1 Row (long press)", requiresLongPress: true)) { item in
+                item.reordering = ItemReordering(sections: .current)
+            }
 
-    private var allRows : [String: DemoItem] = {
-        [
-            "Row 1": DemoItem(text: "Row 1", requiresLongPress: true),
-            "Row 2": DemoItem(text: "Row 2", requiresLongPress: true),
-            "Row 3": DemoItem(text: "Row 3", requiresLongPress: true)
-        ]
-    }()
-
-    func setDragging(_ id: String, dragging: Bool) {
-        var row = allRows[id]
-        row?.dragging = dragging
-        if let row = row {
-            allRows[id] = row
-        }
-    }
-
-    func reset() {
-        allRows.forEach { (key, _) in
-            setDragging(key, dragging: false)
+            section += Item(DemoItem(text: "3,2 Row (long press)", requiresLongPress: true)) { item in
+                item.reordering = ItemReordering(sections: .current)
+            }
         }
     }
 }